Move check-spelling to modules
