+++ modules/check-spelling.js
